# docs/swagger.yaml
basePath: /api/v1
definitions:
  dtos.CreateCategoryRequest:
    properties:
      categoryName:
        example: nigger comedy
        type: string
    type: object
  dtos.CreateStreamHistoryRequest:
    properties:
      hostPrincipalId:
        example: user123
        type: string
      streamHistoryStreamId:
        example: stream123
        type: string
      videoUrl:
        example: supabase storage video
        type: string
    type: object
  dtos.CreateStreamHistoryResponse:
    properties:
      duration:
        type: integer
      hostPrincipalId:
        type: string
      streamHistoryId:
        type: string
      streamHistoryStreamId:
        type: string
      videoUrl:
        type: string
    type: object
  dtos.CreateStreamInfoRequest:
    properties:
      categoryId:
        example: "123"
        type: string
      hostPrincipalId:
        example: "123"
        type: string
      title:
        example: My First Stream!!
        type: string
    type: object
  dtos.CreateStreamingResponse:
    properties:
      createAt:
        type: string
      hostPrincipalId:
        type: string
      isActive:
        type: boolean
      streamCategoryName:
        type: string
      streamId:
        type: string
      thumbnail:
        type: string
      title:
        type: string
    type: object
  dtos.CreateViewerHistoryRequest:
    properties:
      viewerHistoryPrincipalID:
        example: user123
        type: string
      viewerHistoryStreamHistoryID:
        example: stream123
        type: string
    type: object
  dtos.CreateViewerHistoryResponse:
    properties:
      viewerHistoryID:
        type: string
      viewerHistoryPrincipalID:
        type: string
      viewerHistoryStreamHistoryID:
        type: string
    type: object
  dtos.GetAllStreamHistoryResponse:
    properties:
      categoryName:
        type: string
      duration:
        type: integer
      hostPrincipalID:
        type: string
      messages:
        items:
          $ref: '#/definitions/dtos.MessageResponse'
        type: array
      streamHistoryID:
        type: string
      streamHistoryStreamID:
        type: string
      thumbnail:
        type: string
      title:
        type: string
      totalView:
        type: integer
      videoUrl:
        type: string
    type: object
  dtos.MessageResponse:
    properties:
      content:
        type: string
      createdAt:
        type: string
      messageID:
        type: string
      senderID:
        type: string
    type: object
  dtos.StreamInfoResponse:
    properties:
      categoryName:
        type: string
      hostPrincipalId:
        type: string
      title:
        type: string
    type: object
  dtos.StreamResponse:
    properties:
      categoryName:
        type: string
      createdAt:
        type: string
      hostPrincipalID:
        type: string
      isActive:
        type: boolean
      messages:
        items:
          $ref: '#/definitions/dtos.MessageResponse'
        type: array
      streamId:
        type: string
      thumbnailURL:
        type: string
      title:
        type: string
      viewerCount:
        type: integer
    type: object
  dtos.UpdateStreamActiveStatusRequest:
    properties:
      hostPrincipalID:
        example: stream123
        type: string
    type: object
  models.Category:
    properties:
      categoryID:
        type: string
      categoryName:
        type: string
    type: object
  models.Message:
    properties:
      content:
        type: string
      createdAt:
        type: string
      messageID:
        type: string
      messagePrincipalID:
        type: string
      streamID:
        type: string
    type: object
info:
  contact:
    email: Jacksontpa7@gmail.com
    name: Jackson API Support
  description: API documentation for the streaming live chat backend.
  license:
    name: MIT
  title: Streaming Backend API
  version: "1.0"
paths:
  /category/create-category:
    post:
      consumes:
      - application/json
      description: Create Category
      parameters:
      - description: Category Name
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create Category
      tags:
      - Category
  /category/get-all-category:
    get:
      consumes:
      - application/json
      description: Retrieve all categories from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all categories
      tags:
      - Category
  /messages/{stream_id}:
    get:
      description: Retrieve all chat messages by Stream ID
      parameters:
      - description: Stream ID
        in: path
        name: stream_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Message'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all messages for a stream
      tags:
      - messages
  /stream-history/all-stream:
    get:
      consumes:
      - application/json
      description: Get all stream history by streamerID
      parameters:
      - description: Host Principal ID
        in: query
        name: hostPrincipalID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/dtos.GetAllStreamHistoryResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all stream history by streamerID
      tags:
      - Stream History
  /stream-history/by-id:
    get:
      consumes:
      - application/json
      description: Get stream history by streamHistoryID
      parameters:
      - description: Stream History ID
        in: query
        name: streamHistoryID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.GetAllStreamHistoryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get stream history by streamHistoryID
      tags:
      - Stream History
  /stream-history/create:
    post:
      consumes:
      - application/json
      description: Create Stream History
      parameters:
      - description: Stream History
        in: body
        name: CreateStreamHistory
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateStreamHistoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.CreateStreamHistoryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create Stream History
      tags:
      - Stream History
  /stream-info/:
    put:
      consumes:
      - application/json
      description: Create or update stream info
      parameters:
      - description: DTO
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateStreamInfoRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create or update stream info
      tags:
      - StreamInfo
  /stream-info/{hostPrincipalID}:
    get:
      consumes:
      - application/json
      description: get stream info by user id
      parameters:
      - description: Host Principal ID
        in: path
        name: hostPrincipalID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.StreamInfoResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: get stream info by user id
      tags:
      - StreamInfo
  /streams/all-active-stream:
    get:
      consumes:
      - application/json
      description: Get all active stream
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.StreamResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all active stream
      tags:
      - Stream
  /streams/by-stream-id:
    get:
      consumes:
      - application/json
      description: Get active stream by streamID
      parameters:
      - description: Stream ID
        in: query
        name: streamID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.StreamResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get active stream by streamID
      tags:
      - Stream
  /streams/create-stream:
    post:
      consumes:
      - multipart/form-data
      description: Create Stream with thumbnail upload
      parameters:
      - description: Host Principal ID
        in: formData
        name: hostPrincipalId
        required: true
        type: string
      - description: Stream title
        in: formData
        name: title
        required: true
        type: string
      - description: Stream Category ID
        in: formData
        name: streamCategoryId
        required: true
        type: string
      - description: Thumbnail file
        in: formData
        name: thumbnail
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.CreateStreamingResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create Stream
      tags:
      - Stream
  /streams/stop-stream:
    post:
      consumes:
      - application/json
      description: Stop active stream
      parameters:
      - description: DTO
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateStreamActiveStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.StreamResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Stop active stream
      tags:
      - Stream
  /viewer-history/create:
    post:
      consumes:
      - application/json
      description: Create Viewer History
      parameters:
      - description: Viewer History
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateViewerHistoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.CreateViewerHistoryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create Viewer History
      tags:
      - Viewer History
swagger: "2.0"